Refactor EditMenu.jsx (#29)
* #25 extract validation functions and image upload functions

* #25 add prop types for ImageUpload.jsx

* #25 extract form to EditMenuForm

* #25 ignore image changes in static folder for menu
