+++ Project/Frontend/src/EditMenu.jsx
@@ -1,17 +1,23 @@
 import { useEffect, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import {
-  TextField,
   Button,
   Box,
   Typography,
-  Select,
-  MenuItem,
-  InputLabel,
-  FormControl,
   Snackbar,
   Alert,
 } from "@mui/material";
+
+import {
+  validateName,
+  validateDescription,
+  validatePrice,  
+  validateCalories, 
+  validateCategory
+} from "./EditMenu/menuValidations";
+
+import ImageUpload from "./EditMenu/ImageUpload";
+import EditMenuForm from "./EditMenu/EditMenuForm";
 
 function EditMenu() {
   // state variables 
@@ -30,10 +36,6 @@ function EditMenu() {
     category: false
   }); // validation errors for form fields 
 
-  // Options for allergens and dietary restrictions 
-  const ALLERGENS_OPTIONS = ["GLUTEN", "DAIRY", "PEANUTS", "SHELLFISH"];
-  const DIETARY_RESTRICTIONS_OPTIONS = ["VEGETARIAN", "VEGAN", "HALAL"];
-
   // fetch the menu item details when component loads
   useEffect(() => {
     fetch(`http://localhost:8080/MenuItems/get/${id}`)
@@ -45,32 +47,8 @@ function EditMenu() {
       .catch((error) => console.error("Error fetching menu item:", error));
   }, [id]);
 
-  const validateName = (name) => {
-    // Name should be 3-50 characters, only letters, spaces, and hyphens
-    return /^[A-Za-z\s-]{3,50}$/.test(name);
-  };
-
-  const validateDescription = (description) => {
-    // Description should be 10-250 characters
-    return description.length >= 10 && description.length <= 250;
-  };
-
-  const validatePrice = (price) => {
-    // Price should be a positive number between 0 and 21
-    return price > 0 && price < 21;
-  };
-
-  const validateCalories = (calories) => {
-    // Calories should be between 0 and 2000
-    return calories >= 0 && calories <= 2000;
-  };
-
-  const validateCategory = (category) => {
-    // Category should be between 0 and 3
-    return category >= 0 && category <= 3;
-  };
-
   // handle changes to the form fields 
+  // validate functions are imported from EditMenu/ 
   const handleChange = (event) => {
     const { name, value } = event.target;
     
@@ -91,10 +69,9 @@ function EditMenu() {
   };
 
   // handle image upload 
-  const handleImageUpload = async (event) => {
-    const file = event.target.files[0];
+  const handleImageUpload = async (file) => {
     if (!file) return;
-
+    
     const formData = new FormData();
     formData.append("image", file);
 
@@ -114,19 +91,6 @@ function EditMenu() {
     } catch (error) {
       console.error("Error uploading image:", error);
     }
-  };
-
-  // handle drag-and-drop image upload 
-  const handleDrop = (event) => {
-    event.preventDefault();
-    const file = event.dataTransfer.files[0];
-    if (file) {
-      handleImageUpload({ target: { files: [file] } });
-    }
-  };
-
-  const handleDragOver = (event) => {
-    event.preventDefault();
   };
 
   // handle form submission 
@@ -209,175 +173,17 @@ function EditMenu() {
       <Typography variant="h4" sx={{ fontWeight: "bold", marginBottom: 3 }}>
         Edit Menu Item
       </Typography>
-
+    
       {/* Form */}
-      <form onSubmit={handleSubmit}>
-        {/* Name Field */}
-        <TextField
-          label="Name"
-          name="name"
-          value={menuItem.name}
-          onChange={handleChange}
-          fullWidth
-          required
-          error={errors.name}
-          helperText={errors.name ? "Name must be 3-50 characters, only letters, spaces, and hyphens" : ""}
-          sx={textFieldStyles}
-        />
-
-        {/* Description Field */}
-        <TextField
-          label="Description"
-          name="description"
-          value={menuItem.description}
-          onChange={handleChange}
-          fullWidth
-          required
-          multiline
-          rows={3}
-          error={errors.description}
-          helperText={errors.description ? "Description must be 10-250 characters" : ""}
-          sx={textFieldStyles}
-        />
-
-        {/* Price Field */}
-        <TextField
-          label="Price"
-          name="price"
-          type="number"
-          value={menuItem.price}
-          onChange={handleChange}
-          fullWidth
-          required
-          error={errors.price}
-          helperText={errors.price ? "Price must be between 0 and 21" : ""}
-          inputProps={{ step: "0.01", min: "0", max: "1000" }}
-          sx={textFieldStyles}
-        />
-
-        {/* Image Upload */}
-        <Box
-          sx={{
-            border: "2px dashed white",
-            padding: "20px",
-            textAlign: "center",
-            cursor: "pointer",
-            mb: 2,
-          }}
-          onDragOver={handleDragOver}
-          onDrop={handleDrop}
-        >
-          <Typography sx={{ color: "white" }}>
-            Drag & drop an image here, or click to select
-          </Typography>
-          <input
-            type="file"
-            accept="image/*"
-            onChange={handleImageUpload}
-            style={{ display: "none" }}
-            id="file-upload"
-          />
-          <label htmlFor="file-upload">
-            <Button variant="outlined" component="span" sx={{ mt: 1, color: "white", borderColor: "white" }}>
-              Select File
-            </Button>
-          </label>
-        </Box>
-        {imagePath && <img src={imagePath} alt="Menu Item" width="150px" />}
-
-        {/* Calories Field */}
-        <TextField
-          label="Calories"
-          name="calories"
-          type="number"
-          value={menuItem.calories}
-          onChange={handleChange}
-          fullWidth
-          required
-          error={errors.calories}
-          helperText={errors.calories ? "Calories must be between 0 and 2000" : ""}
-          inputProps={{ min: "0", max: "2000" }}
-          sx={textFieldStyles}
-        />
-
-        {/* Category Field */}
-        <TextField
-          label="Category"
-          name="category"
-          type="number"
-          value={menuItem.category}
-          onChange={handleChange}
-          fullWidth
-          required
-          error={errors.category}
-          helperText={errors.category ? "Category must be between 0 and 3" : ""}
-          inputProps={{ min: "0", max: "3" }}
-          sx={textFieldStyles}
-        />
-
-        {/* Allergens Multi-Select */}
-        <FormControl fullWidth sx={{ mb: 2 }}>
-          <InputLabel sx={{ color: "white" }}>Allergens</InputLabel>
-          <Select
-            name="allergens"
-            multiple
-            value={menuItem.allergens}
-            onChange={handleChange}
-            sx={selectStyles}
-          >
-            {ALLERGENS_OPTIONS.map((option) => (
-              <MenuItem key={option} value={option}>
-                {option}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-
-        {/* Dietary Restrictions Multi-Select */}
-        <FormControl fullWidth sx={{ mb: 2 }}>
-          <InputLabel sx={{ color: "white" }}>Dietary Restrictions</InputLabel>
-          <Select
-            name="dietaryRestrictions"
-            multiple
-            value={menuItem.dietaryRestrictions}
-            onChange={handleChange}
-            sx={selectStyles}
-          >
-            {DIETARY_RESTRICTIONS_OPTIONS.map((option) => (
-              <MenuItem key={option} value={option}>
-                {option}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-
-        {/* Availability Dropdown */}
-        <FormControl fullWidth sx={{ mb: 2 }}>
-          <InputLabel sx={{ color: "white" }}>Available</InputLabel>
-          <Select
-            name="available"
-            value={menuItem.available}
-            onChange={handleChange}
-            sx={selectStyles}
-          >
-            <MenuItem value={true}>Available</MenuItem>
-            <MenuItem value={false}>Not Available</MenuItem>
-          </Select>
-        </FormControl>
-
-        {/* Submit Button */}
-        <Button
-          type="submit"
-          variant="contained"
-          sx={{
-            backgroundColor: "#5762d5",
-            color: "white",
-            "&:hover": { backgroundColor: "#4751b3" },
-          }}
-        >
-          Save Changes
-        </Button>
-      </form>
+      <EditMenuForm 
+          imageComponent={
+              <ImageUpload imagePath={imagePath} onImageChange={handleImageUpload}></ImageUpload>
+          }
+          handleSubmit={handleSubmit}
+          menuItem={menuItem}
+          handleChange={handleChange}
+          errors={errors}
+      ></EditMenuForm>
 
       {/* Snackbar for Notifications */}
       <Snackbar open={open} autoHideDuration={6000} onClose={handleClose}>
@@ -388,21 +194,5 @@ function EditMenu() {
     </Box>
   );
 }
-
-const textFieldStyles = {
-  mb: 2,
-  "& .MuiInputBase-input": { color: "white" },
-  "& .MuiInputLabel-root": { color: "white" },
-  "& .MuiOutlinedInput-root": {
-    "& fieldset": { borderColor: "white" },
-    "&:hover fieldset": { borderColor: "white" },
-    "&.Mui-focused fieldset": { borderColor: "white" },
-  },
-};
-
-const selectStyles = {
-  color: "white",
-  "& .MuiOutlinedInput-notchedOutline": { borderColor: "white" },
-};
 
 export default EditMenu;
